+++ config/index.js
@@ -1,0 +1,17 @@
+let config = {
+    local: {
+        mode: 'local',
+        port: 3000
+    },
+    staging: {
+        mode: 'staging',
+        port: 4000
+    },
+    production: {
+        mode: 'production',
+        port: 5000
+    }
+};
+module.exports = function(mode) {
+    return config[mode || process.argv[2] || 'local'] || config.local;
+};
