+++ src/server/src/tests/search.spec.js
@@ -1,0 +1,4 @@
+'user strict';
+
+const chai = require('chai');
+const { assert } = chai;
